// types/Post.ts
export type Post = {
    id: number;
    date: string;
    date_gmt: string;
    guid: {
        rendered: string;
    };
    modified: string;
    modified_gmt: string;
    slug: string;
    status: "publish" | "future" | "draft" | "pending" | "private";
    type: string;
    link: string;
    title: {
        rendered: string;
    };
    content: {
        rendered: string;
        protected: boolean;
    };
    excerpt: {
        rendered: string;
        protected: boolean;
    };
    author: number;
    featured_media: number;
    comment_status: "open" | "closed";
    ping_status: "open" | "closed";
    sticky: boolean;
    template: string;
    format:
        | "standard"
        | "aside"
        | "chat"
        | "gallery"
        | "link"
        | "image"
        | "quote"
        | "status"
        | "video"
        | "audio";
    meta: any;
    categories: number[];
    tags: number[];
};